Add a "One-Key-Halt" function to the remote control, and update my FLYC program

+++ FLYC/motor.h
@@ -1,2 +1,2 @@
 Servo motor[4];
-float Motor[4];
+int Motor[4];
